requirements.txt
# Pflug Law Lead Qualifier - Python Dependencies

# API Clients
requests>=2.31.0
anthropic>=0.39.0
openai>=1.0.0
pyairtable>=2.0.0

# Google APIs (Gmail + Drive)
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Web Dashboard
flask>=3.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type hints (optional but recommended)
types-requests>=2.31.0
